WIP fix specs
